Create Staff and Student models
Create a staff model with a OneToOne field to the User model, also with a boolean flag that signals if the staff member has admin access or not.

Create a student model with a OneToOne field to the User, a class group and a year group.
Add a function to get the full class name for display purposes.