Create UserState model
This model will hold only an email address, and a flag saying whether the email address is of a staff member, or a student.
This model will be checked against when creating users, and the appropriate Staff/Student model will be created.
The database table this model corresponds to will be updated every hour by a django command, that uses the Google App Manager to check email groups and add either staff, or students emails to the table.