Complete post_save function
Get a state instance from the user state table, and populate the appropriate fields in the model to be saved.

This is maybe not the most efficient way to go about this, but it allows there to be a single point of knowledge concerning what a certain email address can or can't do.

TODO: catch users who are not already in the state table.