Redesign class_group models
A single ClassGroup model is not sufficient to represent the entire school, and is tough to work with inside templates (having to pass through context values from YEAR_CHOICES etc instead of retrieving the appropriate model instance and using that).

Model hierarchy should be as follows:
School > Year > Band > Set

This is more representative of what the organisational structure of a school is like, and will be easier to work with inside templates.